move operations AppDbContext into the shared JsonDotNetCoreExample project
having a single db  makes CI easier to configure
