+++ src/Examples/JsonApiDotNetCoreExample/Program.cs
@@ -5,10 +5,7 @@ namespace JsonApiDotNetCoreExample
 {
     public class Program
     {
-        public static void Main(string[] args)
-        {
-            BuildWebHost(args).Run();
-        }
+        public static void Main(string[] args) => BuildWebHost(args).Run();
 
         public static IWebHost BuildWebHost(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
