+++ src/Examples/JsonApiDotNetCoreExample/appsettings.json
@@ -1,14 +1,13 @@
 {
-  "Data": {
-    "DefaultConnection":
-      "Host=localhost;Port=5432;Database=JsonApiDotNetCoreExample;User ID=postgres;Password=password"
-  },
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Trace",
-      "System": "Trace",
-      "Microsoft": "Trace"
+    "Data": {
+        "DefaultConnection": "Host=localhost;Port=5432;Database=JsonApiDotNetCoreExample;User ID=postgres;Password=postgres"
+    },
+    "Logging": {
+        "IncludeScopes": false,
+        "LogLevel": {
+            "Default": "Trace",
+            "System": "Trace",
+            "Microsoft": "Trace"
+        }
     }
-  }
 }
