+++ src/Examples/OperationsExample/OperationsExample.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <ProjectReference Include="../../JsonApiDotNetCore/JsonApiDotNetCore.csproj" />
+    <ProjectReference Include="..\JsonApiDotNetCoreExample\JsonApiDotNetCoreExample.csproj" />
   </ItemGroup>
 
   <ItemGroup>
