+++ src/Examples/OperationsExample/Startup.cs
@@ -1,13 +1,12 @@
 using System;
 using JsonApiDotNetCore.Extensions;
+using JsonApiDotNetCoreExample.Data;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using OperationsExample.Data;
-
 
 namespace OperationsExample
 {
