+++ src/Examples/OperationsExample/appsettings.json
@@ -1,6 +1,6 @@
 {
     "Data": {
-        "DefaultConnection": "Host=localhost;Port=5432;Database=OperationsExample;User ID=postgres;Password=postgres"
+        "DefaultConnection": "Host=localhost;Port=5432;Database=JsonApiDotNetCoreExample;User ID=postgres;Password=postgres"
     },
     "Logging": {
         "IncludeScopes": false,
