+++ test/JsonApiDotNetCoreExampleTests/appsettings.json
@@ -1,14 +1,13 @@
 {
-  "Data": {
-    "DefaultConnection":
-      "Host=localhost;Port=5432;Database=JsonApiDotNetCoreExample;User ID=postgres;Password="
-  },
-  "Logging": {
-    "IncludeScopes": false,
-    "LogLevel": {
-      "Default": "Error",
-      "System": "Information",
-      "Microsoft": "Information"
+    "Data": {
+        "DefaultConnection": "Host=localhost;Port=5432;Database=JsonApiDotNetCoreExample;User ID=postgres;Password=postgres"
+    },
+    "Logging": {
+        "IncludeScopes": false,
+        "LogLevel": {
+            "Default": "Error",
+            "System": "Information",
+            "Microsoft": "Information"
+        }
     }
-  }
 }
